Extract SqlDependencyProcessDispatcherStorage
