+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft.Data.SqlClient.csproj
@@ -101,58 +101,55 @@
   <!-- Contains common items shared between NetFx and NetCore -->
   <ItemGroup>
     <Compile Include="$(CommonSourceRoot)Interop\Common\Sni\SniErrors.cs">
-      <Link>Interop\SniErrors.cs</Link>
+      <Link>Interop\Common\SniErrors.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)Interop\Windows\AdvApi32\AdvApi32.netfx.cs">
-      <Link>Interop\AdvApi32\AdvApi32.netfx.cs</Link>
+      <Link>Interop\Windows\AdvApi32\AdvApi32.netfx.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)Interop\Windows\Kernel32\FileTypes.cs">
-      <Link>Interop\Kernel32\FileTypes.cs</Link>
+      <Link>Interop\Windows\Kernel32\FileTypes.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)Interop\Windows\Kernel32\IoControlCodeAccess.cs">
-      <Link>Interop\Kernel32\IoControlCodeAccess.cs</Link>
+      <Link>Interop\Windows\Kernel32\IoControlCodeAccess.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)Interop\Windows\Kernel32\IoControlTransferType.cs">
-      <Link>Interop\Kernel32\IoControlTransferType.cs</Link>
+      <Link>Interop\Windows\Kernel32\IoControlTransferType.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Kernel32\Kernel32.cs">
-      <Link>Interop\Kernel32\Kernel32.cs</Link>
+      <Link>Interop\Windows\Kernel32\Kernel32.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Kernel32\Kernel32Safe.netfx.cs">
-      <Link>Interop\Kernel32\Kernel32Safe.netfx.cs</Link>
+      <Link>Interop\Windows\Kernel32\Kernel32Safe.netfx.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\CreateDisposition.cs">
-      <Link>Interop\NtDll\CreateDisposition.cs</Link>
+      <Link>Interop\Windows\NtDll\CreateDisposition.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\CreateOptions.cs">
-      <Link>Interop\NtDll\CreateOptions.cs</Link>
+      <Link>Interop\Windows\NtDll\CreateOptions.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\DesiredAccess.cs">
-      <Link>Interop\NtDll\DesiredAccess.cs</Link>
+      <Link>Interop\Windows\NtDll\DesiredAccess.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\FileFullEaInformation.cs">
-      <Link>Interop\NtDll\FileFullEaInformation.cs</Link>
+      <Link>Interop\Windows\NtDll\FileFullEaInformation.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\ImpersonationLevel.cs">
-      <Link>Interop\NtDll\ImpersonaltionLevel.cs</Link>
+      <Link>Interop\Windows\NtDll\ImpersonaltionLevel.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)Interop\Windows\NtDll\IoStatusBlock.cs">
-      <Link>Interop\NtDll\IoStatusBlock.cs</Link>
+      <Link>Interop\Windows\NtDll\IoStatusBlock.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\NtDll.cs">
-      <Link>Interop\NtDll\NtDll.cs</Link>
+      <Link>Interop\Windows\NtDll\NtDll.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\ObjectAttributeFlags.cs">
-      <Link>Interop\NtDll\ObjectAttributes.cs</Link>
+      <Link>Interop\Windows\NtDll\ObjectAttributes.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\ObjectAttributes.cs">
-      <Link>Interop\NtDll\ObjecAttributes.cs</Link>
+      <Link>Interop\Windows\NtDll\ObjecAttributes.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\NtDll\SecurityQualityOfService.cs">
-      <Link>Interop\NtDll\SecurityQualityOfService.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonSourceRoot)Interop\Windows\SystemErrors.cs">
-      <Link>Interop\SystemErrors.cs</Link>
+      <Link>Interop\Windows\NtDll\SecurityQualityOfService.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Secur32\CredHandle.netfx.cs">
       <Link>Interop\Windows\Secur32\CredHandle.netfx.cs</Link>
@@ -197,13 +194,19 @@
       <Link>Interop\Windows\Sni\SniDnsCacheInfo.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Sni\SniError.cs">
-      <Link>Interop\Windows\Sni\SniErro.cs</Link>
+      <Link>Interop\Windows\Sni\SniError.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonSourceRoot)\Interop\Windows\Sni\SqlDependencyProcessDispatcherStorage.netfx.cs">
+      <Link>Interop\Windows\Sni\SqlDependencyProcessDispatcherStorage.netfx.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Sni\TransparentNetworkResolutionMode.cs">
       <Link>Interop\Windows\Sni\TransparentNetworkResolutionMode.cs</Link>
     </Compile>
+    <Compile Include="$(CommonSourceRoot)Interop\Windows\SystemErrors.cs">
+      <Link>Interop\Windows\SystemErrors.cs</Link>
+    </Compile>
     <Compile Include="$(CommonSourceRoot)Interop\Windows\UnicodeString.cs">
-      <Link>Interop\UnicodeString.cs</Link>
+      <Link>Interop\Windows\UnicodeString.cs</Link>
     </Compile>
     <Compile Include="$(CommonSourceRoot)Microsoft\Data\Common\ActivityCorrelator.cs">
       <Link>Microsoft\Data\Common\ActivityCorrelator.cs</Link>
